+++ src/CommunityToolkit.Maui.Core/Extensions/WindowExtensions.android.cs
@@ -21,21 +21,4 @@ public static class AndroidWindowExtensions
 		window.AddFlags(WindowManagerFlags.DrawsSystemBarBackgrounds);
 		return window;
 	}
-
-
-	/// <summary>
-	/// Checks if the specified Android version is supported.
-	/// </summary>
-	/// <param name="version">The Android version to check.</param>
-	/// <returns>True if the version is supported; otherwise, false.</returns>
-	public static bool IsSupported(BuildVersionCodes version)
-	{
-		if (OperatingSystem.IsAndroidVersionAtLeast((int)version))
-		{
-			return true;
-		}
-
-		System.Diagnostics.Trace.WriteLine($"This functionality is not available. Minimum supported API is {version}");
-		return false;
-	}
 }
